+++ app.js
@@ -2,44 +2,38 @@ const parse = require('csv-parser')
 const fs = require('fs')
 const yargs = require('yargs')
 const chalk = require('chalk');
-const { validarFechas, asignarYear, compararFecha, mostrarFecha } = require('./helper.js')
+const hr = require('./helper.js')
 
 
 if(!process.argv[2]){
-    console.log(chalk.cyan.bold('--------------------------------------------------------------------------------------------')+ '\r\n' +
-                chalk.cyan.bold('This app finds all the employees birthdays between two dates at most separated for 365 days.' + '\r\n' + 
-                                'Please provide the required two dates following the next format:') + '\r\n' +
-                chalk.magenta.bold('node app.js addDates --first="mm/dd" --second="mm/dd"')+ '\r\n' + 
-                chalk.cyan.bold('Please consider that the first date takes the cuurent year and if you provide the greater' + '\r\n' + 'date as first, as well as if both dates are the same, the second takes one year later.') +'\r\n'+
-                chalk.cyan.bold('--------------------------------------------------------------------------------------------'))
-}
+    hr.instruCCions()
+} 
 
+const funcionCumpleannos = (a, b) => {
 
-const funcionCumpleannos = (a,b) =>{
-    
-    if(!validarFechas(a,b)){
-        console.log(chalk.red('Please use the required format'))
-        return false
+    if (!hr.validarFechas(a, b)) {
+        console.log(chalk.red('Please use the required format'));
+        return false;
     }
 
-    const { fecha1, fecha2 } = asignarYear(a,b)
-
-    const cumpleanios = []
+    const { fecha1, fecha2 } = hr.asignarYear(a, b);
+    const cumpleanios = [];
 
     fs.createReadStream('mails_y_cumples_03.csv')
         .pipe(parse({
-                delimiter: ','
-            })
+            delimiter: ','
+        })
         )
         .on("data", (dataRow) => {
-            if(compararFecha(fecha1, fecha2, dataRow.cumpleanios.split('-'))){
-                cumpleanios.push(dataRow)
+            if (hr.compararFecha(fecha1, fecha2, dataRow.cumpleanios.split('-'))) {
+                cumpleanios.push(dataRow);
             }
         })
-        .on("end",() => {
-            mostrarFecha(cumpleanios)
-        })
+        .on("end", () => {
+            hr.mostrarFecha(cumpleanios);
+        });
 }
+
 
 yargs.command({
     command: 'addDates',
